refactor: ResultViewPlayer가 Player를 상속하도록 수졍
